Move player/enemy collision code
Moved player/enemy collision code from 'EnemyMovement' script to 'EnemyHealth' script.

Previously, a bug occured where upon collision, player HP wasn't decremented and enemy game object is still being destroyed.

Bug fixed after moving code, probably shifting code updates the frame more accurately to decrement player HP before destroying itself.
