+++ Assets/Prefabs/EnemyPrefab.prefab
@@ -173,7 +173,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5d20cc51a8370d940bef4773d6a27333, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  damage: 0
   moveSpeed: 0.5
   animator: {fileID: 1288170906316595519}
 --- !u!114 &-7339899325233340433
@@ -190,6 +189,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   health: 1
   animator: {fileID: 1288170906316595519}
+  damage: 1
 --- !u!95 &1288170906316595519
 Animator:
   serializedVersion: 5
